move heartbeat logic to probe-agent
