+++ config/manager-probe-agent/manager.yaml
@@ -23,7 +23,7 @@ spec:
         app: probe-agent
     spec:
       containers:
-        - image: kubeprober/probe-agent:v0.1.0-beta1
+        - image: kubeprober/probe-agent:v0.1.0-beta2
           imagePullPolicy: "Always"
           name: probe-agent
           command: ["/bin/sh"]
@@ -123,7 +123,7 @@ spec:
         app: probe-tunnel
     spec:
       containers:
-        - image: kubeprober/probe-tunnel:v0.1.0-beta1
+        - image: kubeprober/probe-tunnel:v0.1.0-beta2
           imagePullPolicy: "Always"
           name: probe-agent
           command: ["/bin/sh"]
