+++ config/manager-probe-master/manager.yaml
@@ -28,7 +28,7 @@ spec:
           env:
             - name: SERVER_SECRET_KEY
               value: your-token-here
-          image: kubeprober/probe-master:v0.1.0-beta1
+          image: kubeprober/probe-master:v0.1.0-beta2
           imagePullPolicy: "Always"
           name: probe-master
           securityContext:
