+++ probers/addon/mysql/main.sh
@@ -7,6 +7,8 @@ MYSQL_PORT=$(kubectl  get cm dice-addons-info -n default -o jsonpath='{.data.MYS
 MYSQL_DATABASE=$(kubectl  get cm dice-addons-info -n default -o jsonpath='{.data.MYSQL_DATABASE}')
 MYSQL_PASSWORD=$(kubectl  get cm dice-addons-info -n default -o jsonpath='{.data.MYSQL_PASSWORD}')
 
+MYSQL_USERNAME=$MYSQL_USERNAME"dbmigration"
+
 function check_mysql() {
   #check mysql connected
   mysql -h $MYSQL_HOST -u $MYSQL_USERNAME  -P $MYSQL_PORT -p$MYSQL_PASSWORD -e "select now()" >/dev/null 2>/dev/null
